# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "qianji",
    "description": "Universal Browser Automation for AI",
    "version": "2.0.0"
  },
  "paths": {
...
    "/browsers/create": {
      "post": {
        "tags": [
          "basic"
        ],
        "summary": "Create Browser",
        "description": "创建新浏览器实例\n\nArgs:\n    profile: 浏览器配置文件名称\n    browser_id: 指定的浏览器ID（可选）\n    idle_timeout: 空闲超时时间（秒），默认使用服务器配置\n    max_lifetime: 最大生命周期（秒），默认使用服务器配置\n    headless: 是否使用无头模式，默认使用服务器配置\n    no_sandbox: 是否禁用沙箱，默认使用服务器配置\n    args: 额外的浏览器启动参数",
        "operationId": "create_browser_browsers_create_post",
        "parameters": [
          {
            "name": "profile",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile"
            }
          },
          {
            "name": "browser_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Browser Id"
            }
          },
          {
            "name": "idle_timeout",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idle Timeout"
            }
          },
          {
            "name": "max_lifetime",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Lifetime"
            }
          },
          {
            "name": "headless",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Headless"
            }
          },
          {
            "name": "no_sandbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "No Sandbox"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Args"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browsers/{browser_id}/close": {
      "post": {
        "tags": [
          "basic"
        ],
        "summary": "Close Browser",
        "description": "关闭指定浏览器实例",
        "operationId": "close_browser_browsers__browser_id__close_post",
        "parameters": [
          {
            "name": "browser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Browser Id"
            }
          },
          {
            "name": "purge_profile",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Purge Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "targetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetid"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "snapshot": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot"
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "GenericResponse",
        "description": "通用响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}